dlt-ingestion-framework/tests/integration/test_end_to_end.py
"""
Integration test: PostgreSQL to DuckDB.

This test requires:
- PostgreSQL database running (Docker recommended)
- Sample data loaded

Run with: pytest -m integration tests/integration/test_postgresql_to_duckdb.py
"""
import pytest
from pathlib import Path
import duckdb

from src.core.orchestrator import IngestionOrchestrator


@pytest.mark.integration
@pytest.mark.skipif(
    not Path('/tmp/test_postgres_available').exists(),
    reason="PostgreSQL not available"
)
class TestPostgreSQLToDuckDB:
    """End-to-end tests for PostgreSQL ingestion."""
    
    @pytest.fixture
    def orchestrate_config(self, temp_dir):
        """Create configuration for integration test."""
        import pandas as pd
        from pathlib import Path
        
        # Create test config
        config_data = [{
            'source_type': 'postgresql',
            'source_name': 'integration_test',
            'table_name': 'test_table',
            'load_type': 'FULL',
            'enabled': 'Y'
        }]
        
        df = pd.DataFrame(config_data)
        excel_path = temp_dir / 'integration_config.xlsx'
        df.to_excel(excel_path, sheet_name='SourceConfig', index=False)
        
        return excel_path
    
    def test_full_load_end_to_end(self, temp_dir, orchestrator_config):
        """Test complete full load pipeline."""
        # Setup DuckDB as destination
        db_path = temp_dir / 'test.duckdb'
        
        # Configure orchestrator
        orchestrator = IngestionOrchestrator(config_dir=temp_dir)
        
        # Execute job
        job = {
            'source_type': 'postgresql',
            'source_name': 'integration_test',
            'table_name': 'test_table',
            'load_type': 'FULL',
            'enabled': 'Y'
        }
        
        # Run ingestion
        # orchestrator.execute_job(job)
        
        # Verify data in DuckDB
        # conn = duckdb.connect(str(db_path))
        # result = conn.execute("SELECT COUNT(*) FROM test_table").fetchone()
        # assert result[0] > 0
        
        # Placeholder for full implementation
        pytest.skip("Integration test requires PostgreSQL setup")
    
    def test_incremental_load_end_to_end(self):
        """Test complete incremental load pipeline."""
        pytest.skip("Requires PostgreSQL with sample data")
    
    def test_schema_evolution_detection(self):
        """Test schema change detection during ingestion."""
        pytest.skip("Requires PostgreSQL with schema changes")


@pytest.mark.integration
class TestOracleToDuckDB:
    """End-to-end tests for Oracle ingestion."""
    
    def test_oracle_full_load(self):
        """Test Oracle full load."""
        pytest.skip("Requires Oracle database")
    
    def test_oracle_number_type_conversion(self):
        """Test Oracle NUMBER type adapter."""
        pytest.skip("Requires Oracle database")


@pytest.mark.integration
class TestRESTAPIIngestion:
    """End-to-end tests for REST API ingestion."""
    
    @pytest.fixture
    def mock_api_server(self):
        """Start mock API server for testing."""
        # Could use responses library or actual mock server
        pass
    
    def test_api_offset_pagination(self):
        """Test offset-based pagination."""
        pytest.skip("Requires mock API server")
    
    def test_api_cursor_pagination(self):
        """Test cursor-based pagination."""
        pytest.skip("Requires mock API server")
    
    def test_api_authentication(self):
        """Test API key authentication."""
        pytest.skip("Requires mock API server")
